requirements.txt
torch
transformers
pytesseract
PyMuPDF
Pillow